backend/cvd-risk-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
scikit-learn==1.6.1  # Changed to match the version your model was trained with
joblib==1.3.2        # Keeping this as is, typically compatible with scikit-learn 1.6.1
numpy>=1.23.2,<2
groq
python-dotenv==1.0.0
alembic==1.12.1
slowapi==0.1.8
pydantic==2.5.0
pydantic-settings==2.0.3
pydantic[email]
fastapi-cache2==0.2.1